requirements.txt
Flask
gunicorn
sklearn
pandas
ipykernel
pyYAML
six
sklearn
evidently
typing
plotly
dill